Create html markup
Created html boilerplate

Link google font "Overpass" and link style.css.

Created a main container with 2 secondary containers to hold each
section. The first contianer will be the only container shown until the
user selects a rating and submits it. Once the form is submitted, the
first container will be hidden and the second container with the thank
you state will now be shown, thanking the user for their selection.

Radio buttons will be used for the users selection of their rating, and
the span in the secondary thank you state will update depending on which
radio value is selected.