Update span tag with user rating selection
Use an event listener on the submit button to submit and get the users
rating selection. Once the button is selected we will prevent the
default action on the form and loop through the radio button names to
see which button has been selected. Then the span tag from the thank-
you-state text content will receive the value of the checked button.